Revert "feat(sdk): Initial support for migrating W&B runs and reports between instances (#5777)"
This reverts commit f3c9a94af907869f1dfcab1ad71c798b5c8db529.
