+++ requirements_dev.txt
@@ -81,4 +81,3 @@ responses
 kfp<2.0.0
 
 mlflow
-polars
