+++ wandb/apis/importers/__init__.py
@@ -1,6 +1,4 @@
 from wandb.util import get_module
 
-from .wandb import WandbParquetImporter, WandbParquetRun  # noqa: F401
-
 if get_module("mlflow"):
     from .mlflow import MlflowImporter, MlflowRun  # noqa: F401
