+++ wandb/cli/cli.py
@@ -2606,10 +2606,7 @@ def importer():
     help="Override default project to import data into",
 )
 def mlflow(mlflow_tracking_uri, target_entity, target_project):
-    from wandb.util import get_module
-
-    if get_module("mlflow"):
-        from wandb.apis.importers import MlflowImporter
+    from wandb.apis.importers import MlflowImporter
 
     importer = MlflowImporter(mlflow_tracking_uri=mlflow_tracking_uri)
     overrides = {
@@ -2617,4 +2614,4 @@ def mlflow(mlflow_tracking_uri, target_entity, target_project):
         "project": target_project,
     }
 
-    importer.import_all_runs(overrides=overrides)
+    importer.import_all_parallel(overrides=overrides)
