+++ wandb/util.py
@@ -4,7 +4,6 @@ import functools
 import gzip
 import importlib
 import importlib.util
-import itertools
 import json
 import logging
 import math
@@ -14,10 +13,8 @@ import platform
 import queue
 import random
 import re
-import secrets
 import shlex
 import socket
-import string
 import sys
 import tarfile
 import tempfile
@@ -45,7 +42,6 @@ from typing import (
     Set,
     TextIO,
     Tuple,
-    TypeVar,
     Union,
 )
 
@@ -66,7 +62,7 @@ if TYPE_CHECKING:
     from wandb.sdk.artifacts.artifact import Artifact
 
 CheckRetryFnType = Callable[[Exception], Union[bool, timedelta]]
-T = TypeVar("T")
+
 
 logger = logging.getLogger(__name__)
 _not_importable = set()
@@ -1757,38 +1753,3 @@ def cast_dictlike_to_dict(d: Dict[str, Any]) -> Dict[str, Any]:
             d[k] = dict(v)
             cast_dictlike_to_dict(d[k])
     return d
-
-
-def remove_keys_with_none_values(
-    d: Union[Dict[str, Any], Any]
-) -> Union[Dict[str, Any], Any]:
-    # otherwise iterrows will create a bunch of ugly charts
-    if not isinstance(d, dict):
-        return d
-
-    if isinstance(d, dict):
-        new_dict = {}
-        for k, v in d.items():
-            new_v = remove_keys_with_none_values(v)
-            if new_v is not None and not (isinstance(new_v, dict) and len(new_v) == 0):
-                new_dict[k] = new_v
-        return new_dict if new_dict else None
-
-
-def batched(n: int, iterable: Iterable[T]) -> Generator[List[T], None, None]:
-    i = iter(iterable)
-    batch = list(itertools.islice(i, n))
-    while batch:
-        yield batch
-        batch = list(itertools.islice(i, n))
-
-
-def random_string(length: int = 12) -> str:
-    """Generate a random string of a given length.
-
-    :param length: Length of the string to generate.
-    :return: Random string.
-    """
-    return "".join(
-        secrets.choice(string.ascii_lowercase + string.digits) for _ in range(length)
-    )
